Leaner ZIOSpecWithKafka
The new ZIOSpecWithKafka (from #528) is very useful for testing zio-kafka applications, however, it is extended with additional tools naming which are not useful for everyone. Even so, it does force you to implement `kafkaPrefix`.
This PR removes the naming extensions back to the zio-kafka-test module.
