+++ zio-kafka-test/src/test/scala/zio/kafka/ZIOKafkaSpec.scala
@@ -1,0 +1,19 @@
+package zio.kafka
+
+import zio.{ Task, ZIO }
+
+import java.util.UUID
+
+trait ZIOKafkaSpec extends ZIOSpecWithKafka {
+
+  def kafkaPrefix: String
+
+  def randomThing(prefix: String): Task[String] =
+    ZIO.attempt(UUID.randomUUID()).map(uuid => s"$prefix-$uuid")
+
+  def randomTopic: Task[String] = randomThing(s"$kafkaPrefix-topic")
+
+  def randomGroup: Task[String] = randomThing(s"$kafkaPrefix-group")
+
+  def randomClient: Task[String] = randomThing(s"$kafkaPrefix-client")
+}
